a search function in combobox is implemented, needs soc
